Implement multiline input
The previous approach of interpolating the contents of the textarea
directly into the JavaScript template literal containing Python code
didn't work because:
1. The string contained newlines,
   so it had to be specified using triple quotes (""").
2. The template literal is indented within the JavaScript code,
   and the whitespace characters used for indentation
   are part of the string; i.e. the whole Python script is indented;
   however, when the textarea contents got interpolated into it,
   the resulting Python code would contain negative indentation,
   something like this:
   ```python
       from docutils.core import publish_string
       publish_string("""foo
   bar""", writer_name='html5').decode("utf-8")
   ```
   ...so Python would complain about invalid syntax.

Passing the string via a variable
(rather than injecting it directly into the Python code)
avoids this issue because the Python code doesn't change shape
depending on the input text.

In any case, it's unadvisable to directly interpolate user input into code,
because it can lead to code injection,
so the new approach is also safer.