Set iframe content via srcdoc
This fixes the iframe reloading bug in Firefox,
where the page would load correctly the first time,
but on subsequent loads the iframe would load the root page nested into it.

Also replace the `<output>` element with the `<iframe>`,
at least while custom parsing of the docutils output is not implemented.
The `<output>` element can be reintroduced later if appropriate.

Finally, rename the JS variables for DOM elements to reflect the element type.